--- a/apps/public-docsite-v9/.storybook/tsconfig.json
+++ b/apps/public-docsite-v9/.storybook/tsconfig.json
@@ -3,6 +3,7 @@
   "compilerOptions": {
     "allowJs": true,
     "checkJs": true,
+    "jsx": "react",
     "types": ["node", "static-assets", "storybook__addons"]
   },
   "include": ["*.js"]