--- a/apps/public-docsite-v9/tsconfig.json
+++ b/apps/public-docsite-v9/tsconfig.json
@@ -7,8 +7,7 @@
     "experimentalDecorators": true,
     "noUnusedLocals": true,
     "strictNullChecks": true,
-    "noImplicitAny": true,
-    "types": ["webpack-env"]
+    "noImplicitAny": true
   },
   "include": [],
   "files": [],