--- a/apps/react-18-tests-v9/tsconfig.react-compat-check.json
+++ b/apps/react-18-tests-v9/tsconfig.react-compat-check.json
@@ -1,7 +1,0 @@
-{
-  "extends": "./tsconfig.app.json",
-  "compilerOptions": {
-    "noEmit": true,
-    "paths": {}
-  }
-}