--- a/apps/ssr-tests/package.json
+++ b/apps/ssr-tests/package.json
@@ -14,13 +14,13 @@
   },
   "license": "MIT",
   "devDependencies": {
-    "@fluentui/react": "*",
-    "@microsoft/load-themed-styles": "^1.10.26",
     "@types/mocha": "7.0.2",
-    "@fluentui/public-docsite-resources": "*",
     "mocha": "7.2.0",
+    "@microsoft/load-themed-styles": "^1.10.26",
     "@fluentui/scripts-tasks": "*",
-    "@fluentui/scripts-webpack": "*"
+    "@fluentui/scripts-webpack": "*",
+    "@fluentui/react": "*",
+    "@fluentui/public-docsite-resources": "*"
   },
   "dependencies": {
     "tslib": "^2.1.0"