--- a/apps/ssr-tests/test/test.js
+++ b/apps/ssr-tests/test/test.js
@@ -32,6 +32,12 @@ describe('Fabric components', () => {
     let links = AppDefinition.examplePages[i].links;
     for (let j = 0; j < links.length; j++) {
       let { key, component } = links[j];
+      if (!key) {
+        throw new Error(`Component key (current value "${key}") is missing for ${component}`);
+      }
+      if (!component) {
+        throw new Error(`Component (current value "${component}") is missing for ${key}`);
+      }
 
       testRender(key, component);
     }
@@ -55,7 +61,7 @@ describe('Utilities', () => {
 /**
  *
  * @param {string} componentName
- * @param {React.FunctionComponent} component
+ * @param {React.ComponentClass | (() => JSX.Element)} component
  */
 function testRender(componentName, component) {
   it(`${componentName} can render in a server environment`, done => {