<!--
Thank you for submitting a pull request!

Please verify that:
* [ ] Code is up-to-date with the `master` branch
* [ ] Your changes are covered by tests (if possible)
* [ ] You've run `yarn change` locally


PR flow tips:
* [ ] Try to start with a Draft PR
* [ ] Once you're ready (ideally the pipeline is passing) promote your PR to Ready for Review. This step will auto-assign reviewers for your PR.
-->

## Previous Behavior

<!-- This is the behavior we have today -->

## New Behavior

- applies new performant `type-check` command to all apps applicable
- normalizes some apps tsconfig setup with ts solution config pattern
- adds TS setup and type checking for ssr-test v8 app

_Applications type-checking perf improvements:_

| app | Before | After / Delta |
|--------|--------|--------|
| vr-tests | 9.37s | 2.82s / 𝚫70% 🚅|
| vr-tests-react-components | 14s | 5s / 𝚫64% 🚅|
| ssr-tests-v9 | 27.8s | 3.41s / 𝚫88% 🚅|
| recipes-react-components | 12.44s | 2.17s / 𝚫83% 🚅| 
| perf-test-react-components |  12s  | 1.33s  / 𝚫89% 🚅 |
| perf-test |  7s  | 1.74s  /  𝚫75%  🚅 |
| perf | 10.26s |  1.73s / 𝚫83% 🚅 |

## Related Issue(s)

<!-- Please link the issue being fixed so it gets closed when this is merged. -->

- Implements partially https://github.com/microsoft/fluentui/issues/30516