@fabricteam 2024-05-20
## 📊 Bundle size report

✅ No changes found
<div id="bundleSizeComment9423"/>